Style `uint` -> `uint256`

+++ src/interfaces/IOracle.sol
@@ -2,5 +2,5 @@
 pragma solidity >=0.5.0;
 
 interface IOracle {
-    function price() external view returns (uint);
+    function price() external view returns (uint256);
 }
